refactor DynamicRelocator to relax reorder constraints
DynamicRelocator now does two steps, in the first it allocates all the
nessesary space in sections such as dynrel, got, etc.
The second step actually writes the reloc to the data/text/etc sections
this way we can do a relayout with all the section sizes known.
makes for much more readable code and allows feeding two sections of
unknown size, which we'll need later for TLS